fix: archtecture of the project
