+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/gabriellji/Documents/study/project2/sprint/src/components/app/App.js":"1","/Users/gabriellji/Documents/study/project2/sprint/src/index.js":"2","/Users/gabriellji/Documents/study/project2/sprint/src/services/commerce-service.js":"3","/Users/gabriellji/Documents/study/project2/sprint/src/components/card/Card.js":"4","/Users/gabriellji/Documents/study/project2/sprint/src/components/button/Button.js":"5","/Users/gabriellji/Documents/study/project2/sprint/src/components/spinner/Spinner.js":"6","/Users/gabriellji/Documents/study/project2/sprint/src/components/app/transform-view-button/TransformBtn.js":"7","/Users/gabriellji/Documents/study/project2/sprint/src/components/shopping-cart/ShoppingCart.js":"8"},{"size":3999,"mtime":1606254334987,"results":"9","hashOfConfig":"10"},{"size":235,"mtime":1606160338249,"results":"11","hashOfConfig":"10"},{"size":1675,"mtime":1606225880266,"results":"12","hashOfConfig":"10"},{"size":801,"mtime":1606249140928,"results":"13","hashOfConfig":"10"},{"size":268,"mtime":1606170278835,"results":"14","hashOfConfig":"10"},{"size":194,"mtime":1606213415207,"results":"15","hashOfConfig":"10"},{"size":395,"mtime":1606238363410,"results":"16","hashOfConfig":"10"},{"size":326,"mtime":1606252297148,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1tnrtt5",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/gabriellji/Documents/study/project2/sprint/src/components/app/App.js",[],"/Users/gabriellji/Documents/study/project2/sprint/src/index.js",[],"/Users/gabriellji/Documents/study/project2/sprint/src/services/commerce-service.js",[],"/Users/gabriellji/Documents/study/project2/sprint/src/components/card/Card.js",[],"/Users/gabriellji/Documents/study/project2/sprint/src/components/button/Button.js",[],"/Users/gabriellji/Documents/study/project2/sprint/src/components/spinner/Spinner.js",[],"/Users/gabriellji/Documents/study/project2/sprint/src/components/app/transform-view-button/TransformBtn.js",[],"/Users/gabriellji/Documents/study/project2/sprint/src/components/shopping-cart/ShoppingCart.js",[]]
+[{"/Users/gabriellji/Documents/study/project2/sprint/src/components/app/App.js":"1","/Users/gabriellji/Documents/study/project2/sprint/src/index.js":"2","/Users/gabriellji/Documents/study/project2/sprint/src/services/commerce-service.js":"3","/Users/gabriellji/Documents/study/project2/sprint/src/components/card/Card.js":"4","/Users/gabriellji/Documents/study/project2/sprint/src/components/button/Button.js":"5","/Users/gabriellji/Documents/study/project2/sprint/src/components/spinner/Spinner.js":"6","/Users/gabriellji/Documents/study/project2/sprint/src/components/app/transform-view-button/TransformBtn.js":"7","/Users/gabriellji/Documents/study/project2/sprint/src/components/shopping-cart/ShoppingCart.js":"8"},{"size":183,"mtime":1606291443971,"results":"9","hashOfConfig":"10"},{"size":235,"mtime":1606160338249,"results":"11","hashOfConfig":"10"},{"size":1675,"mtime":1606225880266,"results":"12","hashOfConfig":"10"},{"size":801,"mtime":1606249140928,"results":"13","hashOfConfig":"10"},{"size":268,"mtime":1606170278835,"results":"14","hashOfConfig":"10"},{"size":194,"mtime":1606213415207,"results":"15","hashOfConfig":"10"},{"size":395,"mtime":1606238363410,"results":"16","hashOfConfig":"10"},{"size":326,"mtime":1606252297148,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1tnrtt5",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/gabriellji/Documents/study/project2/sprint/src/components/app/App.js",[],"/Users/gabriellji/Documents/study/project2/sprint/src/index.js",[],"/Users/gabriellji/Documents/study/project2/sprint/src/services/commerce-service.js",[],"/Users/gabriellji/Documents/study/project2/sprint/src/components/card/Card.js",[],"/Users/gabriellji/Documents/study/project2/sprint/src/components/button/Button.js",[],"/Users/gabriellji/Documents/study/project2/sprint/src/components/spinner/Spinner.js",[],"/Users/gabriellji/Documents/study/project2/sprint/src/components/app/transform-view-button/TransformBtn.js",[],"/Users/gabriellji/Documents/study/project2/sprint/src/components/shopping-cart/ShoppingCart.js",[]]
